+++ net/dll/seal/relinkeys.h
@@ -18,21 +18,21 @@ SEALDLL HRESULT SEALCALL RelinKeys_Destroy(void* thisptr);
 
 SEALDLL HRESULT SEALCALL RelinKeys_Set(void* thisptr, void* copy);
 
-SEALDLL HRESULT SEALCALL RelinKeys_Size(void* thisptr, int* size);
+SEALDLL HRESULT SEALCALL RelinKeys_Size(void* thisptr, uint64_t* size);
 
 SEALDLL HRESULT SEALCALL RelinKeys_DBC(void* thisptr, int* dbc);
 
 SEALDLL HRESULT SEALCALL RelinKeys_SetDBC(void* thisptr, int dbc);
 
-SEALDLL HRESULT SEALCALL RelinKeys_GetKeyList(void* thisptr, int index, int* count, void** ciphers);
+SEALDLL HRESULT SEALCALL RelinKeys_GetKeyList(void* thisptr, uint64_t index, uint64_t* count, void** ciphers);
 
-SEALDLL HRESULT SEALCALL RelinKeys_HasKey(void* thisptr, int key_power, bool* has_key);
+SEALDLL HRESULT SEALCALL RelinKeys_HasKey(void* thisptr, uint64_t key_power, bool* has_key);
 
-SEALDLL HRESULT SEALCALL RelinKeys_GetKey(void* thisptr, int key_power, int* count, void** ciphers);
+SEALDLL HRESULT SEALCALL RelinKeys_GetKey(void* thisptr, uint64_t key_power, uint64_t* count, void** ciphers);
 
-SEALDLL HRESULT SEALCALL RelinKeys_ClearDataAndReserve(void* thisptr, int size);
+SEALDLL HRESULT SEALCALL RelinKeys_ClearDataAndReserve(void* thisptr, uint64_t size);
 
-SEALDLL HRESULT SEALCALL RelinKeys_AddKeyList(void* thisptr, int count, void** ciphers);
+SEALDLL HRESULT SEALCALL RelinKeys_AddKeyList(void* thisptr, uint64_t count, void** ciphers);
 
 SEALDLL HRESULT SEALCALL RelinKeys_GetParmsId(void* thisptr, uint64_t* parms_id);
